betweenDates 로직 이관
